fix: настроить зависимости в POM-файлах; refactor: внедрить использование модуля stat-dto в модуль stat-server
