+++ stat-service/pom.xml
@@ -11,6 +11,7 @@
 
     <artifactId>stat-service</artifactId>
     <packaging>pom</packaging>
+
     <modules>
         <module>stat-dto</module>
         <module>stat-server</module>
