+++ stat-service/stat-client/src/main/java/ru/practicum/statclient/BaseClient.java
@@ -1,3 +1,5 @@
+package ru.practicum.statclient;
+
 import java.util.List;
 import java.util.Map;
