+++ stat-service/stat-client/src/main/java/ru/practicum/statclient/StatClient.java
@@ -1,3 +1,5 @@
+package ru.practicum.statclient;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.web.client.RestTemplateBuilder;
@@ -5,6 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.web.util.DefaultUriBuilderFactory;
+import ru.practicum.statdto.RequestHitDto;
 
 import java.util.Map;
