--- stat-service/stat-dto/src/main/java/RequestHitDto.java
@@ -1,17 +1,0 @@
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class RequestHitDto {
-    private String app;
-    private String uri;
-    private String ip;
-    private LocalDateTime created;
-}
