+++ stat-service/stat-dto/src/main/java/ru/practicum/statdto/ViewStatDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.statserver.mapper;
+package ru.practicum.statdto;
 
 import lombok.Data;
