+++ stat-service/stat-server/src/main/java/ru/practicum/statserver/mapper/StatMapper.java
@@ -3,6 +3,8 @@ package ru.practicum.statserver.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
+import ru.practicum.statdto.RequestHitDto;
+import ru.practicum.statdto.ViewStatDto;
 import ru.practicum.statserver.model.Hit;
 
 
@@ -10,7 +12,7 @@ import ru.practicum.statserver.model.Hit;
 public interface StatMapper {
     StatMapper INSTANCE = Mappers.getMapper(StatMapper.class);
 
-    Hit toHit(RequestHitDto2 requestHitDto);
+    Hit toHit(RequestHitDto requestHitDto);
 
     ViewStatDto toViewStatDto(ViewStat viewStat);
 }
